requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
filelock==3.16.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT==0.3.2
Flask-JWT-Extended==4.6.0
fsspec==2024.9.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.24.7
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
setuptools==74.1.3
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
